Add ability to define a debugRegex
This change allows finer control about recipients. If the debug flag is activated, all recipients would be replaced by the debug email address.

Now one is able to set up exclusions by using a new property called `debugRegex`. All recipients matching this regex will *not* be replaced by the debug email address regardless of the debug flag.

This is especially useful for test environments where you want emails to be sent to company email addresses but not others.

`DefaultMandrillService` is getting quite complex. I think, we should refactor it in the future and split it up in multiple classes.